Refactors CLI structure using Cobra
This commit refactors the CLI application structure to
utilize the Cobra library for command-line interface
handling. This change improves the organization and
maintainability of the CLI, making it easier to add
new commands and features in the future.

The main logic is moved into separate command files
(start, stop, status, fetch) within the `cmd` package,
enhancing code modularity. The root command is
defined in `cmd/root.go`, and command execution is
handled through the `Execute` function.
