+++ cmd/status.go
@@ -1,0 +1,15 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+var StatusCmd = &cobra.Command{
+	Use:   "status",
+	Short: "Zeigt den Status der Zeiterfassung",
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("Status der Zeiterfassung anzeigen.")
+	},
+}
